+++ Panels/Header.js
@@ -157,7 +157,6 @@ const HeaderBase = kind({
 		 * The following classes are supported:
 		 *
 		 * * `header` - The root class name
-		 * * `input` - Applied to the `headerInput` element
 		 *
 		 * @type {Object}
 		 * @public
@@ -173,30 +172,6 @@ const HeaderBase = kind({
 		 * @private
 		 */
 		entering: PropTypes.bool,
-
-		/**
-		 * [`Input`]{@link sandstone/Input} element that will replace the `title`.
-		 *
-		 * This is also a [slot]{@link ui/Slottable.Slottable}, so it can be referred
-		 * to as if it were JSX.
-		 *
-		 * Note: Only applies to `type="standard"` headers.
-		 *
-		 * Example
-		 * ```
-		 *  <Header>
-		 *  	<title>Example Header Title</title>
-		 *  	<headerInput>
-		 *  		<Input dismissOnEnter />
-		 *  	</headerInput>
-		 *  	<subtitle>The Adventure Continues</subtitle>
-		 *  </Header>
-		 * ```
-		 *
-		 * @type {Node}
-		 * @deprecated To be removed in 1.0.0-beta.1
-		 */
-		headerInput: PropTypes.node,
 
 		/**
 		 * Sets the "hover" state.
@@ -267,17 +242,6 @@ const HeaderBase = kind({
 		 * @public
 		 */
 		onShowBack: PropTypes.func,
-
-		/**
-		 * Sets the visibility of the input field
-		 *
-		 * This prop must be set to true for the input field to appear.
-		 *
-		 * @type {Boolean}
-		 * @deprecated To be removed in 1.0.0-beta.1
-		 * @public
-		 */
-		showInput: PropTypes.bool,
 
 		/**
 		 * A location for arbitrary elements to be placed above the title
@@ -446,7 +410,6 @@ const HeaderBase = kind({
 		closeButtonBackgroundOpacity,
 		css,
 		direction,
-		headerInput,
 		hover,
 		line,
 		marqueeOn,
@@ -454,7 +417,6 @@ const HeaderBase = kind({
 		noCloseButton,
 		onBack,
 		onClose,
-		showInput,
 		slotAbove,
 		slotAfter,
 		slotBefore,
@@ -494,46 +456,6 @@ const HeaderBase = kind({
 			/>
 		) : null);
 
-		// Create the Title component
-		const titleComponent = (
-			<Heading
-				aria-label={title}
-				size="title"
-				spacing="auto"
-				marqueeOn={marqueeOn}
-				forceDirection={direction}
-				alignment={centered ? 'center' : null}
-				className={css.title}
-			>
-				{title}
-			</Heading>
-		);
-
-		let titleOrInput = titleComponent;
-
-		// If there's a headerInput defined, inject the necessary Input pieces and save that as the titleOrInput variable to be used below.
-		if (headerInput) {
-			deprecate({
-				name: 'sandstone/Panels.Header.headerInput and sandstone/Panels.Header.showInput',
-				until: '1.0.0-beta.1'
-			});
-			titleOrInput = (
-				<div className={css.headerInput}>
-					<Transition duration="short" visible={!!showInput} className={css.inputTransition}>
-						<ComponentOverride
-							component={headerInput}
-							className={css.input}
-							css={css}
-							size="large"
-						/>
-					</Transition>
-					<Transition duration="short" direction="down" visible={!showInput}>
-						{titleComponent}
-					</Transition>
-				</div>
-			);
-		}
-
 		// In wizard type, if one slot is filled, automatically include the other to keep the title balanced.
 		// DEV NOTE: Currently, the width of these is not synced, but can/should be in a future update.
 		const bothBeforeAndAfter = (type === 'wizard' && (slotAfter || slotBefore));
@@ -546,7 +468,17 @@ const HeaderBase = kind({
 						<Cell shrink className={css.slotBefore}>{backButton}{slotBefore}</Cell>
 					) : null}
 					<Cell className={css.titleCell}>
-						{titleOrInput}
+						<Heading
+							aria-label={title}
+							size="title"
+							spacing="auto"
+							marqueeOn={marqueeOn}
+							forceDirection={direction}
+							alignment={centered ? 'center' : null}
+							className={css.title}
+						>
+							{title}
+						</Heading>
 						<Heading
 							size="subtitle"
 							spacing="auto"
@@ -570,7 +502,7 @@ const HeaderBase = kind({
 });
 
 const HeaderDecorator = compose(
-	Slottable({slots: ['headerInput', 'title', 'subtitle', 'slotAbove', 'slotAfter', 'slotBefore']}),
+	Slottable({slots: ['title', 'subtitle', 'slotAbove', 'slotAfter', 'slotBefore']}),
 	Skinnable,
 	Toggleable({prop: 'hover', activate: 'onShowBack', deactivate: 'onHideBack', toggle: null}),
 	WindowEventable({globalNode: 'document', onKeyDown: handleWindowKeyPress})